backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
alembic==1.13.2
httpx==0.27.2
beautifulsoup4==4.12.3
itsdangerous==2.2.0
openai==1.51.2
google-generativeai==0.7.2
python-slugify==8.0.4
pyjwt==2.9.0
apscheduler==3.10.4
loguru==0.7.2
markdown==3.7
python-multipart==0.0.12
jinja2==3.1.4
orjson==3.10.7
tenacity==9.0.0
aiofiles==24.1.0
aiosqlite==0.20.0
